Makes the dappmanager more resilience

+++ build/src/src/calls/removePackage.ts
@@ -1,13 +1,16 @@
 import fs from "fs";
 import * as eventBus from "../eventBus";
 // Modules
-import { dockerComposeDown } from "../modules/docker/dockerCommands";
+import { dockerComposeDown, dockerRm } from "../modules/docker/dockerCommands";
 // External call
 import restartPackageVolumes from "./restartPackageVolumes";
 // Utils
 import * as getPath from "../utils/getPath";
 import shell from "../utils/shell";
 import { RpcHandlerReturn } from "../types";
+import { listContainer } from "../modules/docker/listContainers";
+import Logs from "../logs";
+const logs = Logs(module);
 
 /**
  * Remove package data: docker down + disk files
@@ -17,22 +20,24 @@ import { RpcHandlerReturn } from "../types";
  */
 export default async function removePackage({
   id,
-  deleteVolumes = false
+  deleteVolumes = false,
+  timeout = 10
 }: {
   id: string;
   deleteVolumes?: boolean;
+  timeout?: number;
 }): Promise<RpcHandlerReturn> {
   if (!id) throw Error("kwarg id must be defined");
 
-  const packageRepoDir = getPath.packageRepoDir(id, false);
-  const dockerComposePath = getPath.dockerComposeSmart(id);
-  if (!fs.existsSync(dockerComposePath)) {
-    throw Error(`No docker-compose found: ${dockerComposePath}`);
+  const { name, isCore, packageName: containerName } = await listContainer(id);
+
+  if (isCore || id === "dappmanager.dnp.dappnode.eth") {
+    throw Error("Core packages cannot be cannot be removed");
   }
 
-  if (id.includes("dappmanager.dnp.dappnode.eth")) {
-    throw Error("The installer cannot be removed");
-  }
+  // Only no-cores will
+  const composePath = getPath.dockerCompose(name, false);
+  const packageRepoDir = getPath.packageRepoDir(id, false);
 
   /**
    * [NOTE] Not necessary to close the ports since they will just
@@ -41,10 +46,27 @@ export default async function removePackage({
 
   // Call restartPackageVolumes to safely delete dependant volumes
   if (deleteVolumes) await restartPackageVolumes({ id, doNotRestart: true });
-  // Remove container (and) volumes
-  await dockerComposeDown(dockerComposePath, deleteVolumes);
+  else {
+    /**
+     * If there is no docker-compose, do a docker rm directly
+     * Otherwise, try to do a docker-compose down and if it fails,
+     * log to console and do docker-rm
+     */
+    if (fs.existsSync(composePath))
+      try {
+        await dockerComposeDown(composePath, {
+          volumes: deleteVolumes,
+          timeout
+        });
+      } catch (e) {
+        logs.error(`Error on dockerComposeDown of ${id}: ${e.message}`);
+        await dockerRm(containerName, { volumes: deleteVolumes });
+      }
+    else await dockerRm(containerName, { volumes: deleteVolumes });
+  }
+
   // Remove DNP folder and files
-  await shell(`rm -r ${packageRepoDir}`);
+  if (fs.existsSync(packageRepoDir)) await shell(`rm -r ${packageRepoDir}`);
 
   // Emit packages update
   eventBus.requestPackages.emit();
